Refactored project structure
